bug 1429933: Convert search to use apps.py
Move dbgettest_registration.py to apps.py, so that models will not be
loaded until apps are ready. This is used to describe how filter and
filter group names are extracted from the database, so that they can be
made available for translation.
